Add a service to get description about authenticated service.

+++ server/src/main/java/com/fours/onlineschedulerapi/controller/UserController.java
@@ -2,8 +2,10 @@ package com.fours.onlineschedulerapi.controller;
 
 import com.fours.onlineschedulerapi.model.User;
 import com.fours.onlineschedulerapi.dto.UserDto;
+import com.fours.onlineschedulerapi.service.AuthenticatedUserService;
 import com.fours.onlineschedulerapi.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -16,6 +18,9 @@ public class UserController {
     @Autowired
     private UserService userService;
 
+    @Autowired
+    private AuthenticatedUserService authenticatedUserService;
+
     @PostMapping(value = "/signup")
     public ResponseEntity<?> save(@RequestBody User user) throws EntityExistsException {
         UserDto savedUser = userService.save(user);
@@ -25,9 +30,17 @@ public class UserController {
 
     @DeleteMapping(value = "/{id}")
     public ResponseEntity<?> delete(@PathVariable Long id) throws EntityExistsException {
-        userService.delete(id);
+        if (authenticatedUserService.getAuthorities().contains("COORDINATOR")) {
+            userService.delete(id);
 
-        return ResponseEntity.ok("User deleted successfully.");
+            return ResponseEntity.ok("User deleted successfully.");
+        } else {
+
+            return new ResponseEntity<>(
+                    "You don't have sufficient privilege to complete this request.",
+                    HttpStatus.FORBIDDEN
+            );
+        }
     }
 
     @PutMapping()
